Functionality to not re-install the Edge Hub CA certificate chain unless not installed. This results in improvement in startup latency for smaller memory footprint devices.

+++ edge-hub/src/Microsoft.Azure.Devices.Edge.Hub.Service/scripts/linux/start.sh
@@ -11,6 +11,7 @@ function logtime
 EdgeHubUser=$(printenv EdgeHubUser)
 EdgeModuleHubServerCertificateFile=$(printenv EdgeModuleHubServerCertificateFile)
 EdgeModuleHubServerCAChainCertificateFile=$(printenv EdgeModuleHubServerCAChainCertificateFile)
+EDGEHUB_CA_INSTALLED_FILE="/app/ca_certs_installed.json"
 
 # check if the EdgeAgent supplied the hub with the server certificate and its
 # corresponding signing CA cert
@@ -30,21 +31,28 @@ else
 
     logtime "SSL_CERTIFICATE_PATH=${SSL_CERTIFICATE_PATH}"
     logtime "SSL_CERTIFICATE_NAME=${SSL_CERTIFICATE_NAME}"
-    # copy the CA cert into the ca cert dir
-    command="cp ${EdgeModuleHubServerCAChainCertificateFile} /usr/local/share/ca-certificates/edge-chain-ca.crt"
-    logtime "Executing: ${command}"
-    if ! $command; then
-        logtime "Failed to Copy Edge Chain CA Certificate."
-        exit 1
+
+    # install chain cert only for the first boot
+    if [ ! -f ${EDGEHUB_CA_INSTALLED_FILE} ]; then
+        # copy the CA cert into the ca cert dir
+        command="cp ${EdgeModuleHubServerCAChainCertificateFile} /usr/local/share/ca-certificates/edge-chain-ca.crt"
+        logtime "Executing: ${command}"
+        if ! $command; then
+            logtime "Failed to Copy Edge Chain CA Certificate."
+            exit 1
+        fi
+        # register the newly added CA cert
+        command="update-ca-certificates"
+        logtime "Executing: ${command}"
+        if ! $command; then
+            logtime "Failed to Update CA Certificates."
+            exit 1
+        fi
+        echo " { \"installed_certs\": \"true\" } " > ${EDGEHUB_CA_INSTALLED_FILE}
+        logtime "Certificates installed successfully!"
+    else
+        logtime "Certificates already installed, skipping install!"
     fi
-    # register the newly added CA cert
-    command="update-ca-certificates"
-    logtime "Executing: ${command}"
-    if ! $command; then
-        logtime "Failed to Update CA Certificates."
-        exit 1
-    fi
-    logtime "Certificates installed successfully!"
 fi
 
 # start service
